+++ backend/store/proxy.go
@@ -6,6 +6,7 @@ import (
 	"unsafe"
 
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
+	"github.com/sensu/sensu-go/backend/store/provider"
 	"github.com/sensu/sensu-go/types"
 )
 
@@ -50,6 +51,19 @@ func (e *EventStoreProxy) GetEventByEntityCheck(ctx context.Context, entity, che
 
 func (e *EventStoreProxy) UpdateEvent(ctx context.Context, event *types.Event) (old, new *types.Event, err error) {
 	return e.do().UpdateEvent(ctx, event)
+}
+
+func (e *EventStoreProxy) GetProviderInfo() *provider.Info {
+	p, ok := e.do().(provider.InfoGetter)
+	if ok {
+		return p.GetProviderInfo()
+	}
+	return &provider.Info{
+		TypeMeta: corev2.TypeMeta{
+			Type:       "etcd",
+			APIVersion: "core/v2",
+		},
+	}
 }
 
 type closer interface {